Enforce 32 bit ints for cuda indices

+++ africanus/rime/cuda/tests/test_cuda_predict.py
@@ -45,10 +45,10 @@ def test_cuda_predict_vis(corr_shape, idm, einsum_sig1, einsum_sig2,
     time_idx = np.concatenate([np.full(rows, i+10, dtype=np.int32)
                                for i, rows in enumerate(chunks['rows'])])
 
-    ant1 = np.concatenate([np.random.randint(0, a, rows)
+    ant1 = np.concatenate([np.random.randint(0, a, rows, dtype=np.int32)
                            for rows in chunks['rows']])
 
-    ant2 = np.concatenate([np.random.randint(0, a, rows)
+    ant2 = np.concatenate([np.random.randint(0, a, rows, dtype=np.int32)
                            for rows in chunks['rows']])
 
     assert ant1.size == r
@@ -97,10 +97,10 @@ def test_cuda_apply_gains(corr_shape, idm, einsum_sig1, einsum_sig2, chunks):
     time_idx = np.concatenate([np.full(rows, i+10, dtype=np.int32)
                                for i, rows in enumerate(chunks['rows'])])
 
-    ant1 = np.concatenate([np.random.randint(0, a, rows)
+    ant1 = np.concatenate([np.random.randint(0, a, rows, dtype=np.int32)
                            for rows in chunks['rows']])
 
-    ant2 = np.concatenate([np.random.randint(0, a, rows)
+    ant2 = np.concatenate([np.random.randint(0, a, rows, dtype=np.int32)
                            for rows in chunks['rows']])
 
     assert ant1.size == r
